more on comments
Added the comments frame in landscape.
Added the parsing of the comment parent and changed the sorting
algorithm for comments (e.g. comment always after it's parent and
sorting within comments with the same parent).
Also added the inset mechanism for comments with parents, they are shown
with a small inset, larger while it is lower in the tree.
Removed trailing whitespaces from the spanned which we use for showing
content on screen (simple trim() didn't work on the string because the
html string ended with a </p> which added an extra enter in the end.
Added the notifications option, but don't do anything with it yet.