small fixes
Fix: items in landscape now have a set height. This because the row
height of a row is always determined on the first item in a grid view.
This can cause a second, third, etc item to become bigger than the row.
Fix: IndexOutOfBoundException when opening a post detail really fast.
Fix: IllegalStateException when changing the frame in the homeActivity
just exactly after/while changing orientation, making the frame change
after onSaveInstanceState, could be reproduced sometimes if changing the
orientation while starting up the app.